requirements.txt
pytest : pytest~=9.0.3
selenium>=4.39.0
SQLAlchemy==1.4.50
allure-pytest>=2.15.3